split by files
