+++ src/App/Counter.bt.tsx
@@ -1,0 +1,28 @@
+import userEvent from '@testing-library/user-event';
+import { assert } from 'chai';
+import { Counter } from './Counter';
+import { makeTestSuite, render } from './framework';
+
+const { tests, it } = makeTestSuite();
+
+export { tests };
+
+it('increments', async () => {
+  const screen = render(<Counter start={0} />);
+  const count = screen.getByText('Count: 0');
+
+  const increment = screen.getByRole('button', { name: /Inc/ });
+  await userEvent.click(increment);
+
+  assert.equal(count.innerText, 'Count: 1');
+});
+
+it('decrements', async () => {
+  const screen = render(<Counter start={5} />);
+  const count = screen.getByText('Count: 5');
+
+  const increment = screen.getByText('Dec');
+  await userEvent.click(increment);
+
+  assert.equal(count.innerText, 'Count: 4');
+});
